Fix several issue of mtime related search.
1. If it is an and search, the search result from mtime is not sorted
nor deduplicated, so result may be incomplete or contain duplicate
result. mtimedb may contains multiple mtime of same file.
2. MtimeDB::get should uses quint32 as parameter instead of quint64,
otherwise &mtime may be wrong when passed as search parameter as
quint32.
3. MtimeDB::get didn't call with SET_RANGE thus it always return result
for first key.

REVIEW: 126227

+++ src/engine/mtimedb.h
@@ -44,7 +44,7 @@ public:
     static MDB_dbi open(MDB_txn* txn);
 
     void put(quint32 mtime, quint64 docId);
-    QVector<quint64> get(quint64 mtime);
+    QVector<quint64> get(quint32 mtime);
 
     void del(quint32 mtime, quint64 docId);
